Update README.md
Resized images in category one, included narrative for category two and also reworded some areas though tense should be consistent throughout the page. Included raw narrative for category three and placeholders for visuals in category two. Removed previous repository links to be updated with a new link to the final branch in the repository